// src/app/models/medical-record-out.ts
export interface MedicalRecordOut {
    patientCC: number;
    healthCareProfessionalCC: number;
    vitalSigns: string;
    symptomatology: string;
    injuries: string;
    consultationReason: string;
    diagnosis: string;
    medicalBackground: string;
    interconsultations: string;
    epicrisis: string;
}